Create two unit files for each compilation unit.
This is a transitionary phase while we're switching over to the new
format required by Kythe. The old format encodes the build config (i.e.
the target_os) into the root of every VName, while the new format adds
it as an optional "details" field on the CompilationUnit proto only.

After this CL is submitted, our kzips will contain both formats. In
future, after the frontends accept the new format, I'll remove the
legacy compilation unit.

This is attempt 2 at crrev.com/c/1485631, which had a bug in it where
the BuildDetails message wasn't wrapped in a list.

Bug: 935233
Change-Id: I98af62de5f8dca7ba812f97015e9971cd9634781
Reviewed-on: https://chromium-review.googlesource.com/c/1485631
Commit-Queue: Joey Scarr <jsca@chromium.org>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1496185