[dart] Move test.py variable expansion to test matrix parsing.
This change moves the expansion of the environment variables in
tools/test.py script arguments to the test matrix parsing, where it now
happens every argument for every script instead of just the -n option to
tools/test.py. This change is made to simplify the special cases in the
test matrix, but also to accomodate the use of a tools/disguised_test.py for
the front end team, so they can use test.py but without the new workflow
interfering with it.

While here, remove support for test.py steps without named configuration and
update the example test matrix so it better fits how the builders actually
are.

Change-Id: I487d433360566ae6c24f3ae7048c3f7c6ea5c40b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1496362
Commit-Queue: Jonas Termansen <sortie@google.com>
Reviewed-by: William Hesse <whesse@google.com>