Implement shard retries in the chromium recipe
Shard retries are being added to the chromium recipe. If a test fails,
a merge script can optionally request that a shard be retried. The chromium
recipe will then retry specific shards, and combine the test results
from the original 'with patch' run and the new re-run shards.

See the bug for more details.

Bug: 922152
Change-Id: I539fcfc0c0dc3546c8348f67ff71a1b34cc1faf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1447202
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>