Re-add logic for handling invalid test results in 'with patch'.
This logic was originally written by martiniss@.

The following sequence of events:
 1) 'with patch' produces invalid results
 2) 'retry shards with patch' has a failing test
 3) 'without patch' has a passing test.
 4) 'retry with patch' has a passing test

should count as a FindIt flaky test, with 'retry shards with patch' as the name
of the failing test. Due to some confusion about the nature of the unit test,
this logic was erroneously removed.

Bug: 922152
Change-Id: I56f02aada0cf2597f45a5d422c7ad32d6d24dc1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1509518
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>