Reverse the argument order of assertEquals calls.
assertEquals wants the expected value to be first, for the sake of error
messages. For example, if there are two dicts and the 1st has a key the
2nd doesn't the error will indicate a missing key, not an extra key.

This isn't specified in the docs, and there are conflicting opinions
online. This is the order that I've found to work correctly in practice.

Also s/assertEquals/assertEqual/ while I'm here, as the former is a
deprecated alias for the latter.

Change-Id: I457f4da589c3efac97f6111721b1df4b6f1323a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1521849
Commit-Queue: Owen Rodley <orodley@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>