Force the default cwd for the ToTMacCoverage bot.
Chromium builders that use the chromium recipe and the associated
modules have a bunch of complicated logic to manage doing checkouts
into the 'builder' cache directory so that we reuse checkouts between
builds. In particular, often commands need to set the cache dir as
the current working directory, but the actual logic is much more
complicated than that.

Probably all builders should just set that as the default cwd, and
then we could simplify a bunch of code, but that's way to big a
project to take on at the moment.

Instead, this CL changes just the ToTMacCoverage builder -- which
doesn't use the chromium recipe or much of the other infrastructure --
to do so. This at least "fixes" a problem where one of the chromium
apis confusingly tries to install the mac toolchain as part of
the "runhooks" step (why it does this here, I don't know), but the
install runs in the kitchen directory rather than the builder cache,
making it mostly useless but clobber-y if you're not using the builder cache.

BUG=943088, 943274

Change-Id: I1091a34819fc17e86ff2367a98c1b1c778bab625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1529383
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>