Fix default chromium swarming test placeholders.
This CL is primarily a refactor to emit the correct retcode for swarming collect
steps. This causes more of the chromium_swarming logic to be exercised in tests,
and also changes some of the expectations files.

This CL updates the default chromium swarming placeholders, as well as the
placeholders for several smaller modules including v8, webrtc and iOS.

The changes to the webrtc expectation files are slightly larger, but correct.
The webrtc were written to have the collect step return retcode=1, with no test
results. The test expectations were reflecting retcode=1, but continued to use
the default [successful!] test results.

Change-Id: I042fc94207791c4ef2b15a5a37c1cbbb67a091a1
Bug: 942801
Recipe-Manual-Change: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1529176
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>