Fix canned_isolated_script_output to correctly emit swarming placeholder.
When canned_isolated_script_output() is passed swarming=True, the swarming
summary was generated as a JSON, but was not being set as the placeholder. This
caused tests to incorrectly use the potentially default, or another source of
swarming summary placeholder along with the generated raw_io placeholder.

This CL updates test expectations but has no intended behavior change.

Change-Id: I502949f4620b1d50ae5e852ed866f532915300e1
Bug: 942801
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1534750
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>