[dart] Implement per-CL pre-approval.
This change runs the tools/bots/apply_preapproval.dart script after
downloading the latest results, which applies all pending pre-approvals for
commits that have landed to according to the git history of HEAD.

For try runs, the updated approvals are simply used. For the main CI, the
updated approvals are immediately uploaded to cloud storage and take effect.
The upload only happens if any pre-approvals were applied or expired, which
is why it is the apply_preapproval script that does the uploading. This
solution was picked because it reduces the contention on the approved
results and thus reduces race conditions in practice.

Change-Id: I4a8ddb2aa6b7a4f5355e9a3992c072c333a0a650
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1532560
Reviewed-by: William Hesse <whesse@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>