Add support for optional dimensions to the iOS recipes.
iOS (like Chromium) needs the ability to be able to distribute
tests across a set of configurations, so that we can migrate
from one configuration to another without needing to
change everything.

In the chromium recipe, we added support for an
'optional_dimensions' dict to swarming tasks, that specifies
a set of fallbacks to use if the desired primary configuration
is not availabe within specified intervals.

This CL adds that to the iOS recipe as well, so that we can
use the same functionality and not have to use the old
'trigger multiple dimensions' trigger script to do this in
a much-less-good way.

Bug: 845992
Change-Id: I6a811380dadfabbdc8fa4ae610700b6d3fbc3c64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1540426
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>