Update the filter recipe_module to not use old build properties.
This updates the module to use the fields provided from buildbucket
instead of relying on the old build properties to be set.

In addition, it removes an unneeded test and simplifies some of
the other tests.

We still rely on the 'mastername' property for MB, so that is
unchanged.

Bug: 948474
Change-Id: I4c22f658182d30d80422346cee2e3a7625ca5f3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1549741
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>