Reland "Make invalid test results fail builds"
This reverts commit a50f7e3fb6983d8307a04f27c3f50c41acbfc9e1.

NOTE: If you are going to revert this CL because it's breaking a build,
this could be because this CL uncovered an issue with an existing test.
This CL causes invalid test results to fail builds, when previously they
didn't. If you can fix the broken test rather than reverting this CL,
that's preferrable. If something is urgently broken, this CL
can be reverted cleanly.

Reason for reland: Fixed bad test.

Original change's description:
> Make invalid test results fail builds
>
> Bug: 932316
> Change-Id: If039618bc6dcf6ee17357b9a227df5ddbc0bd7ca
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1524371
> Reviewed-by: Erik Chen <erikchen@chromium.org>
> Reviewed-by: John Budorick <jbudorick@chromium.org>
> Commit-Queue: Stephen Martinis <martiniss@chromium.org>

Change-Id: Icd1f577853dae2ba82d7ce459f366756750245dc
Bug: 932316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1529174
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>