Disable retry with patch by default.
This CL updates several test expectations. The updated tests were expecting
StatusSuccess, but only due to the implicit placeholder for 'retry with patch'.
Now that we're skipping 'retry with patch', all these tests produce
StatusFailure. This is working as intended [and given the way the tests are
named and structured, even expected].

Bug: 946844
Change-Id: I122410ee7ff17a53df502564c733a9fc5df49c26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1551279
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>