Add SkiaRenderer Vulkan FYI bot to GPU.FYI CI recipes
We are looking to add a new bot to the GPU.FYI CI bots. This will be used to
allow us to isolate a new subset of tests for the SkiaRenderer Vulkan project.

This change adds "GPU FYI Linux SkiaRenderer Vulkan (NVIDIA)" to the recipes as
a part of the GPU.FYI CI configuration.

Bug: 946770
Change-Id: I2fa301fddf2310782bcce9cad942fe8bff5eb1da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1544771
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>