[V8] Fix flako cmd line for numfuzz
This passes correct extra-args and variant in the flako command line
for numfuzz cases. The number fuzzer doesn't use variants and the flags
passed to numfuzz can't be used by the test runner to repro later.
Instead we pass the exact flags that were used in a failing test.

Bug: v8:8971
Change-Id: Ie3969ed2db2855b82b653a31198bd86db68a9300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1564474
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>