Refactor ios.trigger() in preparation for migration to chromium_tests.
This CL is intended to be a moderate-sized refactor with no behavior change. The
only major difference is that we no longer try to catch and ignore exceptions in
chromium_swarming.trigger_task(). This reflects the new behavior of the method,
which only raises an exception on unrecoverable error.

This CL contains a slight logic change -- if tasks are not triggered, they are
also not collected.

Everything else is a refactor to explicitly specify state dependencies of
trigger(). This is in preparation of moving the method into a new class [a
subclass of Test].

Tested via LED at:
https://chromium-swarm.appspot.com/task?id=442c77515c140010

Bug:  951182
Change-Id: I964c495c8571be51d595a4ba16891070841ad3cf
Recipe-Manual-Change: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1566573
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>