Fix optional dimension handling of 'host os' for iOS in swarming tasks.
The iOS recipes use 'host os' and 'os' to refer to the host and the device
for on-device tests, and just 'host os' for simulator tests, but the
code in the steps classes for handling optional swarming dimensions
didn't handle that case (they needed to rewrite 'host os' to just 'os').

This fixes the rewriting, but really we need to move to using 'os'
and 'device_os' to match Android (and what swarming expects on our bots).

R=bradhall@chromium.org, lindsayw@chromium.org
BUG=845992, 955856

Change-Id: I959e49e6f705e44144aafe192720719bb750326d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1580444
Reviewed-by: Brad Hall <bradhall@google.com>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>