[code coverage] Skip collecting coverage data for local scripts
There are a number of local isolated scripts such as
check_static_initializers, checkdeps, checkperms etc. that introduce
extra complexitites, meawhile, it's still unclear if there is any value
in collecting code coverage data for them. So, for simplicity, skip
tests that don't run on swarming for now.

Will add the support back if there is a solid use case in the future.

Bug: 944657
Change-Id: Ic0d4139f728a7f3cf9b3faa294b5b4d86d4c4712
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1588615
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>