Clean up chromium_swarming task creation.
This CL is a refactor with no intended behavior change.

Previously, tasks were created by passing around callbacks to callbacks to
callbacks. This was confusing and made it difficult to extend the task creation
machinery to the ios recipes without adding a large amount of complexity.

This CL removes two layers of callbacks. The general mechanism to make a task is
now:
  1) Construct the task with no/minimal parameters. It picks up some state from
     the constructor.
  2) Apply common swarming state.
  3) Apply further customization.

There is one tiny change: previously, there was a recipe-level assertion to
check that extra args were not duplicated when they're added in (1). This is no
longer easily possible, since extra args can be applied at any later state. The
previous assertion did not guard against this either. If this proves important
in the future, we can add machinery for this check, but I'd like to avoid
unnecessary complexity for now.


Change-Id: I28c139db8e91fff13b9525269181cfd5efb8d9a2
Bug: 951182
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1588786
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>