Delete existing metadata if present.
Otherwise changes to this script will linger until the cache is cleared,
or until the source file changes. Even worse, the cache is per-builder,
so it might flap back and forth depending on the builder we're running
on.

To avoid this, instead of skipping the file if metadata is present, we
just wipe the existing metadata. This should ensure consistency
regardless of the existing state.

Bug: 615202
Change-Id: I9bcf7d12f203f83931d6d9f231ca49b61466e47c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1597894
Commit-Queue: Owen Rodley <orodley@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>