Reland "Generate and upload separate coverage data for each target in a build."
Changed the name of the upload directory for per-target metadata so that
it doesn't conflict with where the merged metadata is uploaded.

Previous commit description:

Generate and upload separate coverage data for each target in a build.

This requires us to run llvm-cov on the profdata for each test target.
The code coverage service will aggregate the per-target data to make
reports.

Temporarily skipping this for fuzzers so that the builds don't hit the
20 hr time limit on swarming

Bug: 948885
Change-Id: I17794b6a9aac537b54a2c0c723545d73c37b5f68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1610557
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Sajjad Mirza <sajjadm@chromium.org>