[Test Results] Use canonical format correctly.
canonical.py clearly documents that failures argument should be
a list, but we are passing in a dictionary.

This change clears the way to allow us to change the values of this
dictionary without hitting a lot of noise. I am working on that
change in
https://chromium-review.googlesource.com/c/chromium/tools/build/+/1620673

Bug: 960951
Change-Id: I29002f5566ccbc4661315f7a54f4a9b6cc9f9a25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1622467
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>