chromium_swarming: Return has_valid_results
Swarming now indicates if a swarming task has valid results. A
task can have invalid results if the execution ended improperly. This
can be for a few reasons, including the task timing out or swarming
having an internal failure.

The chromium recipe now uses this information to invalidate test
results if swarming says the test has invalid results. This used
to be done via throwing exceptions, but this behavior was changed
to not do control flow via exceptions, and in the process this
behavior appears to have been lost.

Recipe-Manual-Change: build_limited_scripts_slave
Bug: 965508
Change-Id: Iac2c06dd617345eb22ccdec634276b33f78a5ed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1623552
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>