Display a helpful message in a test summary for invalid results.
The current behavior is that if the 'without patch' step produces
invalid results, we simply don't ignore any failures. This leads to
confusing output where the 'without patch' step will display a list of
failing steps and the 'retry summary' will indicate that those tests
failed with the patch and succeeded without the patch. Now we will
instead indicate that nothing could be ignored.

Change-Id: I17ed35ef03d9db7ce9e174abfd05ada141b96028
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1636529
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>