[dart] Disable workaround for firefox process leaks.
While generally we can carry on because a full results.json has been written
out, sometimes the steps also fail. Unfortunately that means tests can pop
in and out of existence spuriously causing problems with our testing
workflow. Instead of hiding the problem, we should bring it to the front.
That means that whenever we do get results, we get a reliable and full set
of results, and it's more obvious what's wrong with the firefox builders.

Bug: https://github.com/dart-lang/sdk/issues/34158
Change-Id: I68aa6440b3de08015cc77ab24c3d0d493814b82a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1649475
Reviewed-by: William Hesse <whesse@google.com>
Commit-Queue: William Hesse <whesse@google.com>