[dart] Add support for custom test runners
Test matrix steps can now declare a testRunner boolean property. If
that property is "true" the step's script is expected to handle the
--output_directory parameter and create results.json and logs.json files
in the specified directory. Deflaking is not yet support for custom test
runners. The scripts outputs will be treated as if they came from
test.py.

This also fixes a bug in _canned_step that caused the final shard to not
get any step test data.

Change-Id: I696fb29752a7bc5591de8c7144e91ef047f2fb6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1643427
Commit-Queue: Alexander Thomas <athom@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: William Hesse <whesse@google.com>