Write directly into the kzip in package_index.py
Previously we would create a temporary directory with the structure we
wanted and then zip the whole thing up. Now we instead add files into
the kzip directly.

This should hopefully be especially beneficial on the bots, as they run
on GCE with remote disk which has very limited IOPS.

Bug: 873956
Change-Id: I73aa2886282b0fce84870838bd35d5dee88a71e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1659870
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>