[code coverage] Use additional merge script.
Currently, the coverage merge script is not being passed an additional
merge script, because in the default case this is left as None by the
chromium_tests module and populated by the chromium_swarming module
depending on the swarming task type.

This change uses the merge script defaulted to by the swarming task if
not overriden by the test class.

R=martiniss,liaoyuke,sajjadm

Recipe-Manual-Change: build_limited_scripts_slave
Change-Id: I87d838af7904976abc7426244b9e899dffcef6c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1666287
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>