Print Windows crash codes in hex
Failure codes such as STATUS_ACCESS_VIOLATION are easily recognizable
and easily differentiated when printed in hex (0xC0000005) but are
cryptic and conflated when printed as decimal (-1073741819). This change
teaches another of our wrapper scripts to print large negative numbers
as hex so that those skilled in the Windows arts can automatically say
"access violation" or "not access violation." The recipe used was copied
from crrev.com/c/917101.

Before this change:

    exit code (as seen by runtest.py): -1073741502

After this change:

    exit code (as seen by runtest.py): 0xC0000142

0xC0000142 indicates DLL Initialization Failure which was a problem in
this bug and in crbug.com/976161.

Bug: 968440
Change-Id: I23be189daa6a17776dd049da9ba60af04d263879
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1677311
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>