Expose got_cr_revision for substitution
Exposes got_cr_revision as a substitutable value in test specs. This is
because got_revision will be the revision of whatever the "primary" repo
for a builder is, e.g. the V8 revision on V8 builders. Some tests always
require the Chromium revision, so this gives them a way to access it on
all builders.

Bug: 977620
Change-Id: I5149971086b82f30b6d19ce2aebd02955f92665f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1682774
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>