[emscripten-releases] Use JSON description of build and test steps
Instead of hardcoding them in the recipe, read the steps from a file
checked into the emscripten-releases repo.

This change goes along with
https://chromium-review.googlesource.com/c/emscripten-releases/+/1684393

It allows adding new tests much more easily without changes to the recipe itself.
It's currently a simple change doesn't change any recipe capabilities or behavior.

In the long run it's possible we might want more interesting behavior such as changing
the step presentation or more-dynamic determination of which tests to run. In that case
we might want to put the recipe code directly in the releases repo (I'm not sure what
would be involved in that). But for the near future the requirements will stay simple
and this is a good solution, as it avoids CLs against the recipe code and allows e.g.
running trybots on proposed build/stest step changes.

Change-Id: Iafe76b52337373e67b2196a731801688fa8f0f0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1685402
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>