Move chromium celab_version inside the recipe.
I initially moved some of the recipe properties inside the bot config
file as properties because it was nice to have them defined in the repo,
but one of the disadvantages to doing that in chromium is that we need
a signoff from the infra team each time we want to edit it.

I expect us to update this version a couple more times this quarter until things
stabilize, so let's move this property back into the recipe.

I also updated the version of the binaries to contain changes in:
  https://crrev.com/c/1684713/ (cel_ctl upload)
  https://crrev.com/c/1684045/ (cel_ui_agent)

Finally, I'd eventually like to get rid of this property completely and
rely on a single source of truth for both the vpython wheel (test framework)
and the celab binaries since we frequently (though not all the time) need to
update them together.

I logged https://crbug.com/977332 to track this future work.

Change-Id: I53a515cc8a1a673e66343efee929c5a931912e6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1686292
Reviewed-by: Fei Ling <feiling@chromium.org>
Commit-Queue: Mathieu Binette <mbinette@google.com>