[V8] Make it possible to use all supported metric types
The new argument metric_type is intentionally required as it is important that
new users think about which metric type they need and pick one correctly. I have
also renamed the method back to send_value as otherwise we'd have to duplicate
the default parameter values and documentation for each metric type.

This also changes path for infra-python package to infra_python, since dashes
are not allowed in cache names in buildbucket and it is less error-prone when
both cache name and path are the same.

Finally, uploading to ts_mon is rarely relevant to end users, thus we mark the
step as infra such that it becomes purple when something fails.

R=ddoman@chromium.org, nodir@chromium.org, tmrts@chromium.org

Change-Id: Ie074f06dc9a67c7e6d9e2fb90b23321fdd1a2380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1688797
Reviewed-by: Scott Lee <ddoman@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Sergiy Belozorov <sergiyb@chromium.org>