Remove celab_version to use on chromium bots from recipe.
Many changes require changes in both the celab binaries and the python test
framework and having to edit the recipe every time is inconvenient.
We always want to keep those in sync, even if we don't technically *have*
to, because each pair is tested/shipped together out of the CELab repo.

Validation:
https://ci.chromium.org/swarming/task/45fbba65a5ad7f10

Bug: 977332
Change-Id: I09addb34a5644e4192ca626aefb25ca4ea42b034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1699221
Reviewed-by: Fei Ling <feiling@chromium.org>
Commit-Queue: Mathieu Binette <mbinette@google.com>