[V8] Reland: Migrate away from the deprecated Milo APIs
In additional to the original CL, this implements a mechanism for passing
revision range information from parent to the triggered builders. This is
necessary because triggered builders do not receive original Gitiles triggers
from scheduler, but receive Buildbucket triggers with properties. We use these
properties to pass on oldest and newest revision, which is sufficient for the
bisect.


R=machenbach@chromium.org, tmrts@chromium.org

Bug: 945840
Change-Id: Ic5d2a97533b6d446959be36c9c24d3dcb5726417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1699822
Commit-Queue: Sergiy Belozorov <sergiyb@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>