Mark gtest shards as invalid if they were terminated
The json test results format contains a key called `interrupted`,
which results in the test results being considered invalid.
This CL introduces a similar capability to the GTest test results
parser, relying on the 'CAUGHT_TERMINATION_SIGNAL' global flag
which is added to the GTest output json when the test launcher
catches a SIGTERM.

This will result in GTest swarming test shards which run for a
whole hours and then time out -- in which case the Swarming task
sends SIGTERM to the testing process -- not being retried as though
they had simply failed.

Bug: 990058
Change-Id: I391e877ed03322edb3b5a1138754a61c3c23e525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1731137
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>