Kill cocoapods/android SDK/xcode
These components are not used by any tests
on LUCI currently, and take significant time to
install.

For the Android SDK, we should move to using
the same CIPD packages the engine uses when we
actually have a test that needs it on here.

Xcode should be trivial to add back when needed.
It would be nice to come up with a better solution for Cocoapods

Change-Id: Ia548509d31949ffd25578be553005311c09f8aa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1740767
Reviewed-by: Jonah Williams <jonahwilliams@google.com>
Commit-Queue: Dan Field <dnfield@google.com>