Install system Xcode on ToT and upload mac clang bots
An upcoming upstream LLVM change makes it so that an iOS SDK is required
to build the clang/llvm ios runtimes that we bundle.

Chromium's hermetic Xcode doesn't have an iOS SDK, so try to use system Xcode
on the tot bots instead.

This also unblocks removing full Xcode from Chromium's hooks since that
was only there because the clang bots still needed it.

Bug: 976593,985971
Change-Id: I9452b347c334ec6a452b3263dcee2cfcef401629
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1744467
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>