Gatekeeper: exclude some WebRTC builders from default config
In [1] gatekeeper is going to use autogenerated list of builders to gate on.
This list includes some builders that run on WebRTC CI waterfall but are not
intended to cause tree closing. In this CL they are added to
"excluded_builders".

[1] https://chromium-review.googlesource.com/c/chromium/tools/build/+/1742917

Change-Id: Ie2b58867a708a47b77b72f8cd7fc9cd9b533c818
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1756483
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Artem Titarenko <artit@chromium.org>