[dart] Add automatic bisection to flutter triple-head builder
On failure, the recipe will start a bisection process by triggering
additional builds in the commit range between the failed and the
previous build.

If any of the steps in the build fails, bisection starts these steps:
* find previous build
* find commit range between this and the previous build
* schedule a new build

The new build receives the commit in the middle of the commit
range as input commit and the remaining newer and older commits in the
bisect_newer and bisect_older properties. The reason for the failure is
also passed in the bisect_reason properties.

Builds that have bisect_reason set, will:
* build with a checkout of their input commit.
* if successful, the root cause is a newer commit.
* if failing with same reason, the root cause is an older commit.
* if failing for a different reason, there are two failures to pin down.

In the last case, the recipe schedules two additional builds to bisect
in both directions. The original bisect continues in the newer commits,
and the bisect for the new failure is started in the older commits.

https://github.com/dart-lang/sdk/issues/37939

Change-Id: Id69bc91021697d40678b6ef75bb00ec3dfb2da67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1758566
Commit-Queue: Alexander Thomas <athom@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>