Use the summary_markdown feature of recipes results to communicate some helpful Chrome OS specific information about the build. The two things added here are:
1) showing cbb_config, which corresponds to indicating the builder name
2) parsing buildset, to produce a link to the master

This markdown is only displayed on failure as the recipe engine
currently doesn't process summary_markdown on success. Note that this
only attempts to add the summary markdown on the relatively normal exits
of StepFailures being propagated out. Communicating for other types of
exceptions propagating out isn't as clean so we don't bother here.

TEST=./recipes.py test train
BUG=None

Change-Id: Id2046adc99401c25a59f9c21ba4853b16b76bb02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1763305
Commit-Queue: David Burger <dburger@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>