recipes: add a tricium_analyze recipe
This CL introduces a recipe that invokes clang-tidy on changed sources,
and emits clang-tidy's complaints as tricium nits.

At the moment, the intent is to hook this up to Chromium and see what
happens. There's currently special logic to make this bot a nop unless
the person uploading the CL is me, so _hopefully_ I can test this
without disrupting anyone else.

Part of the intent of tricium_analyze is to make it easy to add further
linters that may need the full source tree in the future (e.g.,
Fuchsia's does various formatting things, `pylint`, ...). The hope is
that this is structured in a way that adding these checkers is
straightforward.

The split between tricium_analyze.py and tricium_clang_tidy.py was
mostly born out of convenience and testability. Happy to move code
between the two if people are strongly opinionated one way or another.
Please note that part of the intent of tricium_clang_tidy.py *is* to
have a bot that runs clang-tidy against *everything* in Chromium.
That'll be a special, separate bot, though (mostly for the toolchain
folks to use to vet clang-tidy).

(Primary review was initially done in http://crrev.com/c/1809792)

Bug: 633856
Change-Id: I1dbc4c5b91ce01186986a48bca28c265d2be4823
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1827643
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>