[V8] Show perf-dashboard upload failures
This makes the overall build fail if uploading compile-time
measurements to the perf dashboard fails. Previously, the steps
stayed silently green on error, which led to errors being hidden for
too long.

This moves the measurements to the end of the recipe steps, since the
order is not important, but we want the other steps like triggering
to always have a chance to run.

This is prepared by:
https://crrev.com/c/1839956

Bug: v8:9422
Change-Id: Icb0bbedb108f1d7911ffd28d264fd5b0d77e9a18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1839955
Reviewed-by: Tamer Tas <tmrts@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>