Add support for a 'tests' properties in the chromium recipe
This property allows a chromium tryjob to run only certain tests. This
will be used to allow developers to more easily debug their changes
on different platforms in our infrastructure.

This also adds a new chromium_trybot_experimental recipe, to make it
clear that this is an experimental new feature.

Change-Id: I0bf67bbca2a62fa3b575920100aad7b706db3b5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1845596
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>