Add support for triggering jobs with names not equal to builder name.
Branch CI/CQ is adding multiple builders with the same name in different
buckets which means the job names cannot always be equal to the builder
name.

A property is being added to the chromium_tests module to enable
builders to opt-in to having the jobs they trigger match the expected
format of job names generated by lucicfg for multiple builders with the
same name.

Bug: 1013636
Change-Id: I6ba655e6f5f58f2ceda84b16dc9e657c3907b4ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1856293
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>