Don't pass project=None to schedule_request
The `project` argument to `api.buildbucket.schedule_request` now
defaults to an `INHERIT` sentinel, rather than None, instructing
buildbucket to use the same project for the child build as for the
parent. After a soft transition is complete, None will no longer be
accepted as the value for `project`, so `INHERIT` must be used in places
where the child build should inherit the project from the parent.

More info:
https://crrev.com/950046182d106b95a9686c6c9b168f53da83be57

Change-Id: Id7058da926d834f2a361c8cc588f5336cfe6ff23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1859576
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>