Pull completed_ts and duration into variables
There's a validity check for 'duration' in the shard dict through
if shard.get('duration'). The API for .get states it returns None
if the key is not present. However, within the if block, a
KeyError is being thrown, caused by shard['duration']. Pulling the
value out into a variable before consuming to ensure consistency.

Change-Id: Iea87c778be9d46e34034588ebdc40cfa82b15eeb
Bug: 1013756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1863542
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>