Catapult: Update Recipe for Dashboard Only Testing
We add a property that allows us to control the subset of the tests
we're running in the bots, in preparation for isolating the Dashboard
and Pinpoint tests from the rest of the Catapult tests.

Changes in the remote step will be made in the project hosted files
which generate the build steps using a script.

This change also replaces the owners for the catapult owners to reflect
the new team that owns the recipe testing infrastructure.

R=abennetts@google.com, tandrii@google.com

Bug: chromium:985223
Change-Id: I2089aa676f3f4ab05afe14b63718665b42a83a7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1935583
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>